requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
Django==4.2.9
exceptiongroup==1.2.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.24.4
openai==1.10.0
packaging==23.2
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
pyparsing==3.1.1
pytesseract==0.3.10
pytube==15.0.0
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
